Curso att 2.0

+++ src/main/java/br/com/pocketapps/RankIt/servicies/CursosService.java
@@ -10,7 +10,10 @@ import br.com.pocketapps.RankIt.servicies.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -57,42 +60,61 @@ public class CursosService {
         return new CursosDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!cursosRepository.existsById(id)) {
+            throw new ResourceNotFoundException(("Recurso não encontrado"));
+        }
         try {
             cursosRepository.deleteById(id);
-        } catch (Exception e) {
-            LOGGER.error("Erro ao deletar o curso com ID: " + id, e);
-            throw e;
+        } catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Falha de integridade referencial");
         }
     }
 
     @Transactional
-    public CursosDTO update(CursosDTO dto, Long id) {
-        Cursos entity = convertToEntity(dto);
-        entity.setId(id);
+    public CursosDTO update(Long id, CursosDTO dto) {
+        Cursos entity = cursosRepository.getReferenceById(id);
+        convertToEntityiD(dto, id);
         try {
             entity = cursosRepository.save(entity);
-            LOGGER.info("Curso atualizado com sucesso");
+            System.out.println("Atualizado com sucesso");
+        } catch (EmptyResultDataAccessException e) {
+            RuntimeException exception = new RuntimeException("Erro ao atualizar: ID não encontrado");
         } catch (Exception e) {
-            LOGGER.error("Erro ao atualizar o curso", e);
-            throw e;
+            RuntimeException exception = new RuntimeException("Erro ao atualizar");
+            exception.printStackTrace();
         }
         return new CursosDTO(entity);
     }
 
+    private Cursos convertToEntityiD(CursosDTO dto, Long id) {
+        Cursos entity = cursosRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Curso não encontrado com ID: " + id));
+        return populateCursosFromDTO(dto, entity);
+    }
+
     private Cursos convertToEntity(CursosDTO dto) {
         Cursos entity = new Cursos();
+        return populateCursosFromDTO(dto, entity);
+    }
+
+
+    private Cursos populateCursosFromDTO(CursosDTO dto, Cursos entity) {
         entity.setAuthor(dto.getAuthor());
         entity.setNome(dto.getNome());
+        entity.setNumeroDeAulas(dto.getNumeroDeAulas());
         entity.setNota(dto.getNota());
         entity.setDuracaoTotalCurso(dto.getDuracaoTotalCurso());
         entity.setPreco(dto.getPreco());
         entity.setDescricao(dto.getDescricao());
+        entity.getCategorias().clear();
         for (CategoriaDTO catDTO : dto.getCategorias()) {
             Categoria cat = categoriaRepository.getReferenceById(catDTO.getId());
             entity.getCategorias().add(cat);
         }
         return entity;
     }
+
+
 }
